upgrade to 5.2
not yet finished though

+++ modules/Account/Jobs/Membership/AcceptMembership.php
@@ -1,12 +1,12 @@
 <?php namespace Modules\Account\Jobs\Membership;
 
 use App\Jobs\Job;
-use Illuminate\Contracts\Bus\SelfHandling;
+
 use Illuminate\Foundation\Bus\DispatchesJobs;
 use Modules\Account\MembershipInvitation;
 use Modules\Account\MembershipOwner;
 
-class AcceptMembership extends Job implements SelfHandling
+class AcceptMembership extends Job
 {
 
     use DispatchesJobs;
@@ -25,10 +25,7 @@ class AcceptMembership extends Job implements SelfHandling
     {
         $account = $this->invitation->account;
 
-        if ($this->dispatchFromArray(NewMembership::class, [
-            'account' => $account,
-            'member'  => $this->member
-        ])
+        if ($this->dispatch(new NewMembership($account, $this->member))
         ) {
             return $this->invitation->delete();
         }
